refactor: move open telemetry processor to separated folder
